无人机物体箱定位系统（Flask端云）/server_side/requirements.txt
Flask>=2.2.0
Flask-CORS>=3.0.10
websockets>=10.0
python-socketio>=5.7.0
eventlet>=0.33.0
python-dotenv>=0.19.0
requests>=2.28.0
numpy>=1.21.0
Pillow>=9.0.0
pycryptodome>=3.20.0
sympy>=1.12
